start fix bugging in start algorithm add DataSupplierDTO add teacherDisplay and ClassDisplay remain print the map

+++ EvolutionaryLogic/src/models/FitnessTerminateRule.java
@@ -3,9 +3,9 @@ package models;
 import evolutinary.EvolutionarySystem;
 
 public class FitnessTerminateRule extends TerminateRule {
-    private final int fitness;
+    private final double fitness;
 
-    public FitnessTerminateRule(int fitness){
+    public FitnessTerminateRule(double fitness){
         super(EvolutionarySystem.TerminateRules.ByFitness);
         this.fitness = fitness;
     }
